MetadataExtractor/requirements.txt
pyexiftool==0.5.6
exifread==3.0.0
Pillow==8.4.0
PyPDF2==3.0.1
python-docx==0.8.11
argparse==1.4.0
python-magic==0.4.27
python-magic-bin==0.4.14
chardet==5.2.0
piexif==1.1.3